Disable unnecessary rollback and evict process
When using the object storage interface to write data to ChubaoFS, in
order to ensure the consistency of semantics and object storage,only if
all data is written successfully will affect the existing data, ChubaoFS
will carry out a series of complex metadata operation logic.
During the operation of overwriting the existing file, the file data
will be written to a brand new Inode, forming an invisible temporary
data. When the data is completely written successfully, the new data
will be effective by updating the existing Dentry, Otherwise, temporary
data will be deleted to avoid wasting space.
In the operation of updating the existing Dentry, the operation is
successful but the operation initiating node thinks that the operation
failed due to the timeout of the response.
In this case, the file has Dentry but no Inode. The file can be indexed,
but its data cannot be read or written.

So temporarily disable the Evict command to track this problem.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>